Part 2, also fixed an error in part 1 that didn't occur with my implementation
Theoretically not a difficult problem...

+++ prep_new_day.py
@@ -33,4 +33,5 @@ if __name__ == '__main__':
         os.remove(item)
     
     # delete steps-folder from visu gen
-    shutil.rmtree(os.getcwd() + "/steps")
+    if(os.path.isdir(os.getcwd() + "/steps")):
+        shutil.rmtree(os.getcwd() + "/steps")
